data_pipeline/airflow_project/requirements.txt
alembic==1.7.1
apache-airflow==1.10.12
apispec==1.3.3
appnope==0.1.2
argcomplete==1.12.3
attrs==19.3.0
Babel==2.9.1
backcall==0.2.0
boto3==1.7.84
botocore==1.10.84
cached-property==1.5.2
cattrs==1.0.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.37
cryptography==3.4.8
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.4
dnspython==2.1.0
docutils==0.17.1
email-validator==1.1.3
Flask==1.1.4
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.14
Flask-WTF==0.14.3
funcsigs==1.0.2
future==0.18.2
graphviz==0.17
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
install==1.3.4
ipython==7.16.1
ipython-genutils==0.2.0
iso8601==0.1.16
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
json-merge-patch==0.2
jsonschema==3.2.0
lazy-object-proxy==1.6.0
lockfile==0.12.2
Mako==1.1.5
Markdown==2.6.11
MarkupSafe==2.0.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
natsort==7.1.1
numpy==1.19.5
pandas==1.1.5
parso==0.8.2
pendulum==1.4.4
pexpect==4.8.0
pickleshare==0.7.5
prison==0.2.1
prompt-toolkit==3.0.20
psutil==5.8.0
psycopg2-binary==2.7.4
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
PyJWT==1.7.1
pyrsistent==0.18.0
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
requests==2.26.0
s3transfer==0.1.13
setproctitle==1.2.2
six==1.16.0
SQLAlchemy==1.3.23
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.37.8
tabulate==0.8.9
tenacity==4.12.0
text-unidecode==1.3
thrift==0.13.0
traitlets==4.3.3
typing==3.7.4.3
typing-extensions==3.10.0.2
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==0.16.1
WTForms==2.3.3
zipp==3.5.0
zope.deprecation==4.4.0
